# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Sermo Backend API 서버입니다.",
        "title": "Sermo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/chatbot": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "현재 인증된 사용자의 모든 채팅봇 목록을 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatbot"
                ],
                "summary": "사용자별 채팅봇 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chatbot.ChatbotListResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "새로운 채팅봇을 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatbot"
                ],
                "summary": "채팅봇 생성",
                "parameters": [
                    {
                        "description": "채팅봇 생성 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chatbot.CreateChatbotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/chatbot.CreateChatbotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/chatbot/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 채팅봇 ID로 채팅봇 정보를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatbot"
                ],
                "summary": "채팅봇 ID로 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "채팅봇 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chatbot.ChatbotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 채팅봇 ID로 채팅봇 내용을 수정합니다. UUID와 UserID는 변경할 수 없습니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatbot"
                ],
                "summary": "채팅봇 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "채팅봇 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "수정할 채팅봇 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chatbot.UpdateChatbotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chatbot.UpdateChatbotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 채팅봇 ID로 채팅봇을 삭제합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chatbot"
                ],
                "summary": "채팅봇 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "채팅봇 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chatbot.DeleteChatbotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chatbot.ChatbotListResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hashtags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "chatbot.ChatbotResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hashtags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "chatbot.CreateChatbotRequest": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "상세 설명",
                    "type": "string"
                },
                "gender": {
                    "description": "성별 (male, female, unspecified)",
                    "type": "string"
                },
                "hashtags": {
                    "description": "해시태그 배열",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image_id": {
                    "description": "사진 ID",
                    "type": "string"
                },
                "name": {
                    "description": "채팅봇 이름 (3-100자)",
                    "type": "string"
                }
            }
        },
        "chatbot.CreateChatbotResponse": {
            "type": "object",
            "properties": {
                "chatbot_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "chatbot.DeleteChatbotResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "chatbot.UpdateChatbotRequest": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "상세 설명",
                    "type": "string"
                },
                "gender": {
                    "description": "성별 (male, female, unspecified)",
                    "type": "string"
                },
                "hashtags": {
                    "description": "해시태그 배열",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image_id": {
                    "description": "사진 ID",
                    "type": "string"
                },
                "name": {
                    "description": "채팅봇 이름 (3-100자)",
                    "type": "string"
                }
            }
        },
        "chatbot.UpdateChatbotResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 토큰을 사용한 Bearer 인증",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}